+++ sqitch/deploy/table_player.sql
@@ -6,8 +6,8 @@
 BEGIN;
 
 CREATE TABLE trapparty.player (
-  id                 BIGSERIAL PRIMARY KEY,
-  team_id            BIGINT REFERENCES trapparty.team(id),
+  id                 SERIAL PRIMARY KEY,
+  team_id            INT REFERENCES trapparty.team(id),
   name               TEXT NOT NULL CHECK (char_length(name) < 100),
   invitation_code    UUID NOT NULL UNIQUE DEFAULT trapparty.uuid_generate_v1mc(),
   UNIQUE (team_id, invitation_code)
